bugfix in LibRegistry
Temporary library folders were deleted after the workflow where the
libraries were first used is closed. The libraries were assigned
to the workflow so as soon as the workflow was closed all the temporary
files associated to it were deleted. Now the libraries are stored
instead within the user temporary folder.