Optimization of controlled vocabularies in editor node
Vocabularies are loaded on demand. Every vocabulary has been moved to a
separated spreadsheet (one spreadsheet per vocabulary) to avoid the
performance hit of the Excel library used (POI) which loads all the
sheets in the Excel file.

The memory footprint has been greatly reduced as the formerly static
vocabularies are now loaded on demand.