Replace resource files with working directory
Creator node may now include a reference to a working directory
with resource files. These resource files are not copied through
node instances but only the reference to the local path of the
working directory resulting in faster loading times for nodes using
resources.

Since this setting is a string variable it may be configured with
a flow variable and used in the KNIME server as long as the
working directory is recreated in the server.

The tab "Resource files" in creator node is removed. The working
directory field has to be used instead.

The tab "Resource files" in the editor node is removed. Thus there is
not way to included resources files for new models entered with the
editor node. If it is needed an extra panel could be included to
enter the working directory like in the creator node.

The writer node needed no changes since it was already saving the
resources files into the archive. The reader node while it did not
needed changes it is still much slower than the creator node since
it needs to uncompress the resources out of the OMEX archive to
a temporary folder. It is the only node making a deep copy of the
resources. The folder is temporary and removed by KNIME on close. The
only idea I have now to optimize it is to remove the folder when closing
the workflow if possible (need to check if KNIME supports this on
FileUtil).