[Refactor] replace "file.separator" with forward slash
The "file.separator" is replaced in WriterNodeModel.writeArchive with a
forward slash. The separator was used in the path of a CombineArchive
entry. However, COMBINE archive are platform independent and should not
use specific separators from the machine the model was created on.

Currently JSBML supports the backward slash but it may not in the
future.