Replace R download and install with much faster version
We had another verbose install updating the progress bar of the node
which installed every package on a different step. Installing all the
packages in a single step is much faster and has less cost due to the
boilerplate.

For the plyr package it took:
* W/ verbose: 1037794808969ns (1037.794808969s)
* W/ batch: 330943631252ns (330.943631252s)

Also the CPM model tested went from 1h approx to 10min.