Fix https://github.com/SiLeBAT/FSK-Lab/issues/313
The runner node was running the selected simulation from the input model
that had been selected previously in the simulation configurator node
but the selection in the Runner UI was ignored.

After changes:
- This setting may be configured with a flow variable.
- The setting in the runner takes precedence over the selected
simulation
in the model (port object). If the simulation chosen in the runner does
not exist in the model then selected simulation in the model is used.
- If the runner node is not configured then the selected simulation in
the model is used.