Fix package download
The CRAN mirror used earlier (http://cran.us.r-project.org) has been
replaced with RStudio's as it is replicated globally through a CDN and
may be faster in EU.

The folder 'repo' was not matching the 'cran' folder used by
LigRegistry. Actually, the code between LibRegistry and
RConnectionFactory is a little messy. There is duplication of the
folders creation and they are only created once in RConnectionFactory
which runs first. The functionality should be slowly moved to the server
package *de.bund,.bfr.knime.fsklab.r.server*.