Fix #326. Connection error when no downloads are required
First, if all the packages required by a model are installed and there
is no Internet connection the runner node should not fail. Before, the
runner was always querying via LibRegistry the remote CRAN. If a machine
was not connected the runner would fail even when all the packages were
already installed.

Secondly, a meaningful exception is provided when the machine is not
connected to Internet and a package is missing. In this case the model
cannot be executed and the following error message is shown:
`Execute failed: Not connected to Internet. Packages {triangle} could
not be downloaded`