Make JoinerViewValue.joinerRelations an array
The type of JoinerViewValue.joinerRelations is changed from string to array (JoinRelation[]). Now the serialization/deserialization is done by KNIME instead of the JS view which is now simplified. For simplicity an array is used instead of a list which needs Jackson type boilerplate.

Regarding JoinRelation, the source and target parameters are replaced with the IDs of the source and target parameters. This means that instead of the entire parameter metadata, only the ids are saved which is the only information needed in the JS view.