Move parameters to view representation
Remove the first and second model maths (model1Math and model2Math) in the Joiner
ViewValue and save instead the list of parameters of model 1 and model 2. Only the parameters are used in the view and the rest of the model math is unnecessary.  Also since the parameters are static and do not change, they are moved to the view representation.