Fix persistence of metadata in view value
Fix the serialization and deserialization of metadata in the JoinerViewValue. The JSON string was serialized again and loaded as a string. Now, it is loaded and saved from the node settings as a regular JSON string (which is generated in the NodeModel).

The string does not need to be unescaped since it is not serialized twice.