Refactor de.bund.bfr.knime.tests
should be refactored as a fragment project of different host plugins. It
is also defined currently as a plugin, needing to defined all of its
dependencies which is not needed for fragments. It contains tests for:

- de.bund.bfr.knime.fsklab.nodes for the packages
- de.bund.bfr.knime.fsklab.metadata.model for the package
de.bund.bfr.knime.pmm.fskx and class FskMetaData.
- de.bund.bfr.knime.pmm.nodes for the package
de.bund.bfr.knime.pmm.js.common.