Clean deprecated (#627)
* Clean 1.9 javascript components

Editor, simulator and joiner are moved to single folders following the
naming of the node factory including version. Old components (editor and
joiner) are removed.

* Revert "Clean 1.9 javascript components"

This reverts commit b2a74cebff23c8e05b93c2391f0bb34ca2bbfb7c.

* Restore changes without changing deprecated nodes (error)

* Update plugin.xml

* Work in progress

* Fix CodeMirrors in 1.7.2 editor node

Fix the codemirror used in the 1.7.2 editor node. This node using before a small installation of CodeMirror and is now using the full library from de.bund.bfr.knime.js.

* Fix 1.7.2 JS components