RAKIP vocabularies loaded from bundle
The vocabularie were not loaded before straight from the bundle but a second copy included in the classpath as resources. The vocabularies are already included in the vocabularies library but in the Eclipse bundle they do not have easy access as vocabularies.jar is contained in the service bundle (jar in a jar). Because of this, the resources cannot be loaded from the classpath. As a workaround the SQL files used to copied manually to a resources folder named data. These changes extract vocabularies.jar from the bundlle to a temporary file from which the SQL files are read.

Exceptions are implemented for importing the database. Creating the database or its structure fails the import while errors with individual vocabularies do not. In this case the rest of vocabularies are imported.